Add validate script to CEM tools
